decompose util module
