+++ src/util/dsse.ts
@@ -1,0 +1,25 @@
+/*
+Copyright 2022 The Sigstore Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+const PAE_PREFIX = 'DSSEv1';
+
+// DSSE Pre-Authentication Encoding
+export function preAuthEncoding(payloadType: string, payload: Buffer): Buffer {
+  const prefix = Buffer.from(
+    `${PAE_PREFIX} ${payloadType.length} ${payloadType} ${payload.length} `,
+    'ascii'
+  );
+  return Buffer.concat([prefix, payload]);
+}
