+++ src/util/encoding.test.ts
