+++ src/util/encoding.ts
